// frontend/src/lib/api/model/clusterMetricsResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NodeStatsResponse } from "./nodeStatsResponse";

export interface ClusterMetricsResponse {
  clusterId?: string;
  clusterName?: string;
  status?: string;
  timestamp?: string;
  totalNodes?: number;
  runningNodes?: number;
  avgCpuPercent?: number;
  totalMemoryUsage?: number;
  totalMemoryLimit?: number;
  avgMemoryPercent?: number;
  totalNetworkRx?: number;
  totalNetworkTx?: number;
  totalBlockRead?: number;
  totalBlockWrite?: number;
  nodeMetrics?: NodeStatsResponse[];
  queriesPerSecond?: number;
  activeConnections?: number;
  replicationLagSeconds?: number;
}